requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
numpy==1.24.3
opencv-python==4.7.0.72
Pillow==9.5.0
protobuf==4.23.0
py4j==0.10.9.7
PyBoof==0.41
six==1.16.0
transforms3d==0.4.1
